requirements.txt
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
pyyaml>=6.0
requests>=2.31.0
